week 3 Qns 3, 5, 6

+++ ALGORITHMIC_TOOLBOX/week3_greedy_algorithms/3_maximum_advertisement_revenue/DotProduct.java
@@ -16,7 +16,7 @@ public class DotProduct {
 
         long result = 0;
         for (int i = 0; i < x.length; i++) {
-            result += x[i] * y[i];
+            result += ((long)x[i]) * ((long)y[i]);
         }
         return result;
     }
